Fix rdtsc (fixes random slow-downs and hangs) and unix.com in 64-bit. Getting more usable.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1607 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/include/cpu.h
@@ -122,12 +122,12 @@ union dword {
 
 #define SEG2LINEAR(seg)	((void *)  ( ((uintptr_t)(seg)) << 4)  )
 
-typedef unsigned long FAR_PTR;	/* non-normalized seg:off 32 bit DOS pointer */
+typedef unsigned int FAR_PTR;	/* non-normalized seg:off 32 bit DOS pointer */
 typedef struct {
   u_short offset;
   u_short segment;
 } far_t;
-#define MK_FP16(s,o)	((((unsigned long)s) << 16) | (o & 0xffff))
+#define MK_FP16(s,o)		((((unsigned int)s) << 16) | (o & 0xffff))
 #define MK_FP			MK_FP16
 #define FP_OFF16(far_ptr)	((int)far_ptr & 0xffff)
 #define FP_SEG16(far_ptr)	(((unsigned int)far_ptr >> 16) & 0xffff)
